% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_precept_answers}
\alias{get_precept_answers}
\title{Show Precept Answers}
\usage{
get_precept_answers(pnum)
}
\arguments{
\item{pnum}{precept number}
}
\value{
Nothing.
}
\description{
Launches the pdf viewer to show the questions and answers
in a precept.
}
