% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_problemset}
\alias{get_problemset}
\title{Unpack Problem Set by Number}
\usage{
get_problemset(pnum, newname = NULL)
}
\arguments{
\item{pnum}{Number of problem set}

\item{newname}{Your preferred name for the unpacked problem set materials.
(Default: NULL)}
}
\description{
This function finds a problem set number,
unpacks it into your current working directory, and
provides some hints to get going.
}
\details{
By default the handout materials will unpack into a folder called, e.g.
'problemset1' if 1 is the number you provided.

If you'd prefer to unpack the materials under a different name, perhaps
because you've decided to start fresh, or because for whatever reason there's
already a folder with that name where you want to unpack it, use the
\code{newname} argument to asign a new one.  (Note that this only renames the
top folder. The contents are keep their original names.)
}
