% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{package}
\name{pol345}
\alias{pol345}
\alias{pol345-package}
\title{pol345}
\description{
This package contains precept materials for
Marc Ratkowic's version of POL 345.
}
\section{Unpacking Precept Materials}{


Use \code{\link{get_precept}}, with a number to unpzck precept materials
in a local directory.  For example:

\code{get_precept(2)}

This will unpack the second precept materials into a folder called
'precept2` in your current working directory.
(Type \code{\link{getwd}} if you're
not sure where that is.  You can change it using \code{\link{setwd}} or
through the graphical user interface).
}

\section{Starting Again}{

If you want to start again on the same precept, you'll can unpack the
precept materials again under a different name (\code{\link{get_precept}} won't
overwrite an existing folder.)  To do this use the
\code{newname} argument.  If you want your new copy to be called
'precept2-forreal', then use

\code{get_precept(2, newname = "precept2-forreal")}

Provided there's not already a folder of that name in your
current working directory, you'll get a fresh set of precept materials
unpacked there.

If you want to preview the questions in a precept materials without
unpacking it into your local file system, use \code{\link{preview_precept}}.
}

