% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{preview_precept}
\alias{preview_precept}
\title{Preview Precept Materials}
\usage{
preview_precept(pnum)
}
\arguments{
\item{pnum}{precept number}
}
\value{
Nothing.
}
\description{
Launches the pdf viewer to show the questions in some precept materials.
}
